Fix sign comparison warning
Caused a compile error on g++ 6.2.1 with -Wall and -Werror on my Raspberry Pi 2 using Arch Linux for ARM.

+++ afl-plugin/afl-llvm-pass.so.cc
@@ -110,7 +110,7 @@ bool AFLCoverage::runOnModule(Module &M) {
       BasicBlock::iterator IP = BB.getFirstInsertionPt();
       IRBuilder<> IRB(&(*IP));
 
-      if ((random() % 100) >= inst_ratio) continue;
+      if ((unsigned int)(random() % 100) >= inst_ratio) continue;
 
       /* Make up cur_loc */
